% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{date_full}
\alias{date_full}
\title{Convert Dates to Full-Length Day, Month, Year}
\usage{
date_full(date, abbreviate = FALSE, uk.notation = FALSE)
}
\arguments{
\item{date}{A scalar or vector of length n and class character containing
one or more date values in \code{DD-MMM-YY} format.}

\item{abbreviate}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether to convert months to an abbreviated, 3-character format, e.g. "Aug." Defaults
to \code{FALSE}.}

\item{uk.notation}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether to reorder day, month, and year to common UK notation. Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing
long-form, descriptive dates including the day, full or abbreviated month, and 4-digit
year, in either US or UK notation.
}
\description{
\code{date_full} converts date field values in \code{DD-MMM-YY} format to long-form,
descriptive dates, including full-length years in \code{YYYY} format and
full-length months, with the option to "cleanly" abbreviate them. The option to
convert to the UK date format is also available.
}
\details{
Because the warehouse year format is \code{YY}, dates prior to 1930 C.E. will
automatically convert to post-2000 C.E.

Values passed to argument \code{date =} which are not in \code{DD-MMM-YY}
format or contain unrecognized month abbreviations are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
