% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{date_month}
\alias{date_month}
\title{Convert Dates to Calendar Month}
\usage{
date_month(date, abbreviate = FALSE)
}
\arguments{
\item{date}{A scalar or vector of length n and class character containing
one or more date values in \code{DD-MMM-YY} format.}

\item{abbreviate}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether to convert months to an abbreviated, 3-character format, e.g. "Aug." Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing full-length
or abbreviated month names, e.g. "January" or "Jan", respectively.
}
\description{
\code{date_month} converts date field values in \code{DD-MMM-YY} format to full-length
or abbreviated month names.
}
\details{
Values passed to argument \code{date =} which are not in \code{DD-MMM-YY}
format or contain unrecognized month abbreviations are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
