% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{date_month_year}
\alias{date_month_year}
\title{Convert Dates to Full-Length Month & Year}
\usage{
date_month_year(date, sep = " ", abbreviate = FALSE)
}
\arguments{
\item{date}{A scalar or vector of length n and class character containing
one or more date values in \code{DD-MMM-YY} format.}

\item{sep}{A single, quoted string of class character specifying which punctuation,
characters, or metacharacters are used to separate month and year. Defaults to
a single space, or \code{sep = " "}.}

\item{abbreviate}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether to convert months to an abbreviated, 3-character format, e.g. "Aug." Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing full-length
or abbreviated month names, as well as a 4-digit year in \code{YYYY} format,
separated per user-specified delimiters.
}
\description{
\code{date_month_year} converts date field values in \code{DD-MMM-YY} format to full-length
or abbreviated month names and 4-digit years in \code{YYYY} format.
}
\details{
Because the warehouse year format is \code{YY}, dates prior to 1930 C.E. will
automatically convert to post-2000 C.E.

Values passed to argument \code{date =} which are not in \code{DD-MMM-YY}
format or contain unrecognized month abbreviations are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
