% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{date_term}
\alias{date_term}
\title{Convert Dates to Academic Terms}
\usage{
date_term(date, round.up = FALSE)
}
\arguments{
\item{date}{A scalar or vector of length n and class character containing
one or more date values in \code{DD-MMM-YY} format.}

\item{round.up}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether to convert date values to the subsequent academic term. Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing standardized
terms in \code{YYYYMM} format.
}
\description{
\code{date_term} converts date field values in \code{DD-MMM-YY} format to standard
academic terms in \code{YYYYMM} format. Dates falling within each term are assigned
to the existing semester, i.e. "Fall", "Spring", and "Summer" as \code{08},
\code{01}, and \code{05}, respectively. The option to "round up" to the subsequent
semester is also available.
}
\details{
Notably, conversion to academic terms fluidly opens all \code{term*} family functions,
e.g. \code{term_year_fiscal}, in package \code{panthr}.

Because the warehouse year format is \code{YY}, dates prior to 1930 C.E. will
automatically convert to post-2000 C.E.

Values passed to argument \code{date =} which are not in \code{DD-MMM-YY}
format or contain unrecognized month abbreviations are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{term_dash}, \code{term_date}, \code{term_name}, \code{term_season},
\code{term_year_academic}, \code{term_year_calendar}, \code{term_year_fiscal}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
