% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{date_year_calendar}
\alias{date_year_calendar}
\title{Convert Dates to Calendar Year}
\usage{
date_year_calendar(date)
}
\arguments{
\item{date}{A scalar or vector of length n and class character containing
one or more date values in \code{DD-MMM-YY} format.}
}
\value{
A scalar or vector of values of class character containing full-length
calendar years in \code{YYYY} format.
}
\description{
\code{date_year_calendar} converts date field values in \code{DD-MMM-YY} format to full-length
calendar years in \code{YYYY} format.
}
\details{
Because the warehouse year format is \code{YY}, dates prior to 1930 C.E. will
automatically convert to post-2000 C.E.

Values passed to argument \code{date =} which are not in \code{DD-MMM-YY}
format or contain unrecognized month abbreviations are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
