% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_admit}
\alias{decode_admit}
\title{Convert Admission Codes to Full Descriptions}
\usage{
decode_admit(code, clean = FALSE)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more admission codes.}

\item{clean}{A logical value (i.e. \code{TRUE} or \code{FALSE}) indicating
whether descriptions should be converted to a "cleaner", more human-readable format.
Defaults to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing
full-length admission code descriptions. Missing values are preserved.
}
\description{
\code{decode_admit} converts admission code fields into full-length, descriptive
labels by evaluating a scalar or vector of values of class character.
}
\details{
Values containing unrecognized admission codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
