% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_building}
\alias{decode_building}
\title{Convert Building Codes to Full Labels}
\usage{
decode_building(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more building codes.}
}
\value{
A scalar or vector of values of class character containing
full-length building descriptions. Missing values are preserved.
}
\description{
\code{decode_building} converts building code fields into full-length labels
by evaluating a scalar or vector of values of class character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized building codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
