% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_college}
\alias{decode_college}
\title{Convert College Codes to Full or Abbreviated Labels}
\usage{
decode_college(code, short = FALSE)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more college codes, e.g. \code{BU}, \code{AS}, etc.}

\item{short}{A logical value indicating whether labels should be abbreviated.
Useful for tables and graphics in limiting real estate and non-data ink.
Defaults to \code{short = FALSE}.}
}
\value{
A scalar or vector of values of class character containing
full-length or abbreviated college names. Missing values are preserved.
}
\description{
\code{decode_college} converts all \code{*COLLEGE} code fields, e.g. \code{GRAD_COLLEGE},
into full-length labels by evaluating a scalar or vector of values of class
character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized college codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
