% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_country}
\alias{decode_country}
\title{Convert Country Codes to Full Names}
\usage{
decode_country(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more country codes.}
}
\value{
A scalar or vector of values of class character containing
full-length country names. Missing values are preserved.
}
\description{
\code{decode_country} converts country code fields into full-length names
by evaluating a scalar or vector of values of class character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized country codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
