% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_county}
\alias{decode_county}
\title{Convert County Codes to Full Names}
\usage{
decode_county(code, label = FALSE)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more county codes.}

\item{label}{A logical value (i.e. \code{TRUE} or \code{FALSE}) to
indicate whether returned county names should be labeled with "County". Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing
full-length county names Missing values are preserved.
}
\description{
\code{decode_county} converts county code fields into full-length names
by evaluating a scalar or vector of values of class character. It also
provides the option to include a "County" label.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized county codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
