% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_degree}
\alias{decode_degree}
\title{Convert Degree Codes to Full Labels}
\usage{
decode_degree(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more degree codes, e.g. \code{BA}, \code{MPP}, \code{PHD}, etc.}
}
\value{
A scalar or vector of values of class character containing
full-length degree names. Missing values are preserved.
}
\description{
This function converts all \code{*DEGREE} code fields, e.g. \code{GRAD_DEGREE},
into full-length labels by evaluating a scalar or vector of values of class
character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized degree codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{

}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
