% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_department}
\alias{decode_department}
\title{Convert Department Codes to Full Labels}
\usage{
decode_department(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more department codes, e.g. \code{BIO}, \code{PSY}, etc.}
}
\value{
A scalar or vector of values of class character containing
full-length department names. Missing values are preserved.
}
\description{
This function converts all \code{*DEPARTMENT} code fields, e.g. \code{GRAD_DEPARTMENT},
into full-length labels by evaluating a scalar or vector of values of class
character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized department codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{

}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
