% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_ethnicity}
\alias{decode_ethnicity}
\title{Convert Ethnicity Codes to a Single Label}
\usage{
decode_ethnicity(code)
}
\arguments{
\item{code}{A scalar or vector of length n and either class character or numeric
that contain an ethnicity code, i.e. \code{1} or \code{2}.}
}
\value{
A scalar or vector of values of length n and class character
containing a single label describing the ethnicity code. \code{NULL} values are
converted into the character string: \code{Not Reported}.
}
\description{
This function formats the \code{ETHNIC_CODES} field by evaluating a scalar or vector
of values with ethnic codes and missing values and converting them to a single
label. Missing values are converted to \code{Not Reported}.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing ethnicity codes other than \code{1} or \code{2} are coerced to
\code{NA} (missing) rather than converted to \code{Not Reported}. In these
instances, a warning message is thrown. In instances where race codes
are detected, evaluation halts and an error message is thrown.
}
\seealso{
\code{grepl}, \code{table}, \code{paste}, \code{gsub}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
