% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_ethnorace}
\alias{decode_ethnorace}
\title{Convert Race & Ethnicity Codes to Single Labels}
\usage{
decode_ethnorace(race, ethnicity, ethnic.precedence = TRUE)
}
\arguments{
\item{race}{A scalar or vector of length n and class character comprised of
race codes, e.g. \code{B}, \code{Z}, or some permutation of race codes, e.g. \code{WZI}.}

\item{ethnicity}{A scalar or vector of length n and class character or numeric
that contain an ethnicity code, i.e. \code{1} or \code{2}.}

\item{ethnic.precedence}{A logical value indicating whether ethnicity takes
precedence over one or more races. As \code{Hispanic or Latinx} is the only
other ethnicity, any race or permutation of race codes will be overridden
if \code{TRUE}. Defaults to \code{ethnic.precedence = TRUE}.}
}
\value{
A scalar or vector of values of length n and class character
containing a single label describing the combined race and ethnicity code.

\code{NULL} values are converted into the character string: \code{Not Reported}.
}
\description{
This function formats both the \code{RACE_CODES} and \code{ETHNIC_CODES} fields by
evaluating two scalar values or vectors of values of equal length containing
race and ethnicity codes. Missing values for both race and ethnicity, as well
as missing race codes and ethnicity of \code{1}, are converted to \code{Not Reported}.
In instances where more than one race or ethnicity is evaluated, the value is
labeled \code{Two or More Races}.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Ethnicity codes other than \code{1} or \code{2} are coerced to \code{NA} (missing) rather
than converted to \code{Not Reported}. Likewise, unrecognized race codes are
coerced to \code{NA} (missing) values rather than \code{Not Reported}. In both instances,
a warning is thrown.

In instances were race codes are detected in argument \code{ethnicity =} or
ethnicity codes are detected in argument \code{race =}, evaluation is halted and
an error is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
