% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_grade}
\alias{decode_grade}
\title{Convert Grade Codes to Full Descriptions}
\usage{
decode_grade(code, clean = FALSE)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more grade codes.}

\item{clean}{A logical value (i.e. \code{TRUE} or \code{FALSE}) to
indicate whether returned values should be made more human-readable. Defaults
to \code{FALSE}.}
}
\value{
A scalar or vector of values of class character containing
full-length grade descriptions. Missing values are preserved.
}
\description{
\code{decode_grade} converts grade code fields into full-length labels
by evaluating a scalar or vector of values of class character. It also
allows automated "cleaning" of validated values in a more human-readable
format. Due to duplications in the validation data, multiple values may be
returned. This function is currently in development.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized grade codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{
\code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
