% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_major}
\alias{decode_major}
\title{Convert Major Codes to Full Labels}
\usage{
decode_major(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing
one or more major codes, e.g. \code{CJ}, \code{AAS}, \code{LIT}, etc.}
}
\value{
A scalar or vector of values of class character containing
full-length major names. Missing values are preserved.
}
\description{
This function converts all \code{*MAJOR} code fields, e.g. \code{GRAD_MAJOR},
into full-length labels by evaluating a scalar or vector of values of class
character.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing unrecognized major codes are coerced to \code{NA}
(missing) values and a warning message is thrown.
}
\seealso{

}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
