% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_race}
\alias{decode_race}
\title{Convert Race Codes to a Single Label}
\usage{
decode_race(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing race
codes, e.g. \code{W}, \code{B}, \code{Z}, \code{I}, \code{P}, or any permutation containing more than
one code, e.g. \code{ZI}, \code{WBP}, etc.}
}
\value{
A scalar or vector of values of length n and class character
converted into a single label describing the race code. \code{NULL} values are
converted into the character string: \code{Not Reported}.
}
\description{
\code{decode_race} formats the \code{RACE_CODES} field by evaluating a scalar or vector
of values with race codes and missing values and converting them to a single
category. For example, while \code{B} will convert to \code{Black or African American},
\code{BZ} will convert to \code{Two or More Races}. Missing values are converted to
\code{Not Reported}.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing race codes other than \code{W}, \code{B}, \code{Z}, \code{I}, or \code{P} are
coerced to \code{NA} (missing) rather than converted to \code{Not Reported}. In these
instances, a warning message is thrown. In instances where ethnicity codes
are detected, evaluation halts and an error message is thrown.
}
\seealso{
\code{grepl}, \code{table}, \code{paste}, \code{gsub}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
