% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_race_parsed}
\alias{decode_race_parsed}
\title{Convert Race Codes to Multiple Labels}
\usage{
decode_race_parsed(code)
}
\arguments{
\item{code}{A scalar or vector of length n and class character containing race
codes, e.g. \code{W}, \code{B}, \code{Z}, \code{I}, \code{P}, or any permutation containing more than
one code.}
}
\value{
A scalar or vector of values of length n and class character
converted into full-length labels and separated by commas. \code{NULL} values are
converted into the character string: \code{Not Reported}.
}
\description{
This function formats the \code{RACE_CODES} field by evaluating a scalar or vector
of values with race codes and missing values and converting them to all full
race labels, separated by commas. Missing values are converted to \code{Not Reported}.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values containing race codes other than \code{W}, \code{B}, \code{Z}, \code{I}, or \code{P} are
coerced to \code{NA} rather than converted to \code{Not Reported}. In these instances,
a warning message is thrown.
}
\seealso{
\code{grepl}, \code{table}, \code{paste}, \code{gsub}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
