% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{decode_sex}
\alias{decode_sex}
\title{Convert Sex Codes to Descriptive Labels}
\usage{
decode_sex(code)
}
\arguments{
\item{code}{A 1-character \code{SEX} code, which may be a scalar value or vector
of length n comprised of \code{F}, \code{M}, or \code{NULL}.}
}
\value{
A scalar or vector of values converted into full-length, descriptive labels or
\code{Not Reported} if missing.
}
\description{
\code{decode_sex} formats the \code{SEX} field by evaluating a scalar or vector of
values with gender codes and missing values. \code{F} is converted to \code{Female},
\code{M} is converted to \code{Male}, and \code{NULL} values are converted to \code{Not Reported}.
}
\details{
Output may be made categorical using function
\code{factor} or \code{as.factor}. This function also allows for easily
tallying counts using, e.g., function \code{table}.

Values which are not \code{NULL} nor contain \code{F} or \code{M} are coerced to NA.
}
\seealso{
\code{grepl}, \code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
