% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{field_case}
\alias{field_case}
\title{Modify Case & Separators in Field Names}
\usage{
field_case(data, case = "lower", sep = "_", lower.case = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}, \code{matrix}, or other tabular object with
identifiable column headers or field names.}

\item{case}{A single character string in quotations specifying the case in which
column headers or field names will convert. Acceptable parameters include \code{"lower"},
\code{"upper"}, \code{"title"}, \code{"camel"}, \code{"snake"}, \code{"leopard"},
and \code{"kebab"}. Defaults to \code{case = "lower"}.}

\item{sep}{A single character string in quotations specifying the character(s)
used to separate words in column headers or field names. Defaults to a
single space, i.e. \code{sep = " "}.

Conflicting parameters passed to \code{sep =} are overridden with exotic cases,
since they use specific separators by default (e.g. \code{"-"}, \code{"_"}, \code{"."})
or no separator in the case of \code{"camel"}. In such cases, a warning is thrown
alerting the user of the override.}

\item{lower.case}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
all characters in one or more column headers or field names should be lowercase.
Defaults to \code{lower.case = TRUE}.

Conflicting parameters passed to \code{lower.case =} may be overridden when conflicting
with parameters passed to argument \code{case =}, e.g. \code{"title"}, \code{"upper"},
and \code{"camel"}. This argument distinguishes "upper" and "lower" cases,
particularly for exotic cases, e.g. "upper snake case" and "lower snake case".}
}
\value{
A \code{data.frame} object or other tabular data structure with a modified
case and separators in column headers and field names specified by the user.
}
\description{
\code{field_case} is a versatile function that accepts tabular data with column
headers or field names and modifies their case, separators and, for special
cases, initial letter capitalization.
}
\details{
\code{field_case} automatically formats field names to all lowercase letters
and underscore separators (\code{_}). However, it supports both "lower", "upper",
and "title" cases, exotic cases such as "camel", "kebab", "snake", and "leopard",
and custom separators where applicable.

Objects passed to argument \code{data =} that are not tabular or, if
comprised of a single column, without column names will throw an error.
}
\seealso{
\code{gsub}, \code{names}, \code{colnames}, \code{ncol}, \code{tolower},
\code{toupper}, \code{toTitleCase}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
