% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{field_separator}
\alias{field_separator}
\title{Modify Separating Characters in Field Names}
\usage{
field_separator(data, sep = " ")
}
\arguments{
\item{data}{A \code{data.frame}, \code{matrix}, or other tabular object with
identifiable column headers or field names.}

\item{sep}{A single character string in quotations specifying the character(s)
used to separate words in the column headers or field names. Defaults to a
single space, i.e. \code{sep = " "}.}
}
\value{
A \code{data.frame} object or other tabular data structure with modified
separators in column headers, as specified by the user in \code{sep =}.
}
\description{
\code{field_separator} accepts tabular data with column headers or field names
and modifies the separating characters within the column names. Because warehouse
field names are separated by underscores (\code{_}), this replaces underscores
with one or more user-defined characters or metacharacters.
}
\details{
Objects passed to argument \code{data =} that are not tabular or, if
comprised of a single column, without column names will throw an error.
}
\seealso{
\code{gsub}, \code{names}, \code{colnames}, \code{ncol}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
