% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{gpa_discretize}
\alias{gpa_discretize}
\title{Convert GPA Values into Discrete Ranges}
\usage{
gpa_discretize(gpa, range = "thirds", right = FALSE)
}
\arguments{
\item{gpa}{A scalar or vector of length n and class numeric containing
one or more GPA values.}

\item{range}{A single string, in quotations, indicating the range of cut
points. Possible values include \code{"tenths"}, \code{"quarters"}, \code{"thirds"}, and \code{"halves"}.
Defaults to \code{range = "thirds"}.}

\item{right}{A logical value indicating if ranges should be closed on the right
and open on the left (\code{TRUE}). Defaults to \code{right = TRUE}.}
}
\value{
A scalar or vector of clean character strings, according to the
parameter passed to argument \code{range =}, comprised of the lower and upper
bounds in which the value passed to \code{gpa =} falls. Missing values are preserved.
}
\description{
This function optimizes base R function \code{cut} specifically for values
in \code{*GPA} fields, e.g. \code{GRAD_GPA}, by streamlining binning into tenths,
quarters, thirds, and halves, as well as cleaning bin labels.
}
\details{
Values passed to argument \code{gpa =} which exceed 4.33 are coerced to
\code{NA} (missing) values and a warning message is thrown.
}
\seealso{
\code{cut}, \code{gsub}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
