% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{gpa_round}
\alias{gpa_round}
\title{Round GPA Values Directionally}
\usage{
gpa_round(gpa, digits = 1, direction = "none")
}
\arguments{
\item{gpa}{A scalar or vector of length n and class numeric containing
one or more GPA values.}

\item{digits}{A single numeric value indicating quantity of decimal places to
round. E.g. \code{digits = 2} will round to two decimal places. Parameters exceeding
decimal places of values passed in argument \code{gpa =} automatically round the
last digit. Defaults to \code{digits = 1}.}

\item{direction}{A single string, in quotations, indicating directionality.
Possible values include \code{"up"}, \code{"down"}, or \code{"none"}. Defaults to
\code{direction = "none"}.}
}
\value{
A scalar or vector of GPA values of class numeric containing per the
given parameters. Missing values are preserved.
}
\description{
This function enhances the base R \code{round} function to allow directionality
in rounding GPA' fields, e.g. \code{GRAD_GPA}, allowing for conventional
rounding, rounding up, rounding down, and to which decimal place.
}
\seealso{
\code{round}, \code{ceiling}, \code{floor}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
