% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{term_dash}
\alias{term_dash}
\title{Format Term by Separating Year and Month}
\usage{
term_dash(term)
}
\arguments{
\item{term}{A 6-character \code{TERM} code, which may be a scalar value or vector
of length n in \code{YYYYMM} format.}
}
\value{
A scalar \code{TERM} value or vector of \code{TERM} values in \code{YYYY-MM} format.
}
\description{
This basic function formats the \code{TERM} field by separating the 4-digit year
and 2-digit month with a dash. That is, \code{YYYYMM} converts to \code{YYYY-MM}. It
accepts values of both class numeric and character.
}
\seealso{
\code{substr}, \code{paste}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
