% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{term_name}
\alias{term_name}
\title{Format Term as Season & Year}
\usage{
term_name(term, sep = " ")
}
\arguments{
\item{term}{A 6-character \code{TERM} code, which may be a scalar value or vector
of length n in \code{YYYYMM} format.}

\item{sep}{A character string which separates season and year. Input must be
wrapped in quotes, e.g. \code{" "}, \code{"-"}, or \code{", "}.}
}
\value{
A scalar or vector of values of class character combininging season
and year, e.g. \code{Summer 2016}.
}
\description{
\code{term_name} formats the \code{TERM} field by converting it into a string of class
character containing the season or semester and calendar year, e.g.
\code{Fall 2019}. An option is available to customize how season and year are joined.
}
\details{
Output may be made categorical or ordinal nominal using function
\code{factor} or \code{as.factor}, which may specify order.
}
\seealso{
\code{factor}, \code{as.factor}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
