% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{term_season}
\alias{term_season}
\title{Convert Term to Fall, Spring, or Summer}
\usage{
term_season(term)
}
\arguments{
\item{term}{A 6-character \code{TERM} code, which may be a scalar value or vector
of length n in \code{YYYYMM} format.}
}
\value{
A scalar or vector of values of class character indicating \code{Fall},
\code{Summer}, or \code{Spring} according to the term(s) passed.
}
\description{
\code{term_season} formats the \code{TERM} field by converting it into a string of class
character that matches each term.
}
\details{
Output may be made categorical or ordinal nominal using function
\code{factor} or \code{as.factor}, which may specify order.
}
\seealso{
\code{factor}, \code{as.factor}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
