% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{term_separate}
\alias{term_separate}
\title{Format Term by Separating Year and Month}
\usage{
term_separate(term, sep = "-")
}
\arguments{
\item{term}{A 6-character \code{TERM} code, which may be a scalar value or vector
of length n in \code{YYYYMM} format.}

\item{sep}{A quoted string of class character specifying a delimiter between the
term year and month, e.g. \code{sep = "."} for \code{term = "201908"} results in
\code{2019.08}. Defaults to \code{sep = "_"}.}
}
\value{
A scalar \code{TERM} value or vector of \code{TERM} values in
\code{YYYY*MM} delimited format.
}
\description{
\code{term_separate} is a simple function that formats the \code{TERM} field by
separating the 4-digit year and 2-digit month with a user-defined delimiter.
For example, the \code{YYYYMM} format converts to \code{YYYY-MM} by setting
the delimiter to a dash, or \code{"-"}.
}
\seealso{
\code{substr}, \code{paste}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
