% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panthr.r
\name{term_year_fiscal}
\alias{term_year_fiscal}
\title{Convert Term to Fiscal Year}
\usage{
term_year_fiscal(term)
}
\arguments{
\item{term}{A 6-character \code{TERM} code, which may be a scalar value or vector
of length n in \code{YYYYMM} format.}
}
\value{
A scalar or vector of values in \code{YYYY} format with the fiscal year.
}
\description{
This basic function formats the \code{TERM} field by referencing the semester or
season in order to modify and return the fiscal year during which the term
took place.
}
\details{
Output may be made categorical or ordinal nominal using function
\code{factor} or \code{as.factor}, which may specify order. This function
also allows for easily tallying counts using, e.g., function \code{table}.
}
\seealso{
\code{substr}, \code{table}
}
\author{
Jamison R. Crawford, Institutional Research Associate, Georgia State University
}
